fix: second task QT add long double

+++ src/calculator.h
@@ -3,14 +3,14 @@
 #include <string>
 #include <optional>
 
-using Number = double;
+using Number = long double;
 
 class Calculator {
 public:
     Calculator();
 
-    void Set(Number n);           
-    Number GetNumber() const;     
+    void Set(Number n);           // Установить текущее число
+    Number GetNumber() const;     // Получить текущее число
 
     void Add(Number n);           // Сложение
     void Sub(Number n);           // Вычитание
@@ -18,8 +18,8 @@ public:
     void Mul(Number n);           // Умножение
     void Pow(Number n);           // Возведение в степень
 
-    std::string GetNumberRepr() const;
+    std::string GetNumberRepr() const; // Получить строковое представление числа
 
 private:
-    Number current_number_;
+    Number current_number_;       // Текущее число
 };
